third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-020-expected.txt


FAIL CSS Test: @viewport constrained - extended width and height for non-auto zoom. assert_equals: expected 1600 but got 800
